1629 - AI update datatypes

+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1704,10 +1704,10 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
         Cols.temp_location_sum: pl.Int64,
     }
     test_job_group_percentage_schema = {
-        JobGroupLabels.direct_care: pl.Float64,
-        JobGroupLabels.managers: pl.Float64,
-        JobGroupLabels.regulated_professions: pl.Float64,
-        JobGroupLabels.other: pl.Float64,
+        JobGroupLabels.direct_care: pl.Float32,
+        JobGroupLabels.managers: pl.Float32,
+        JobGroupLabels.regulated_professions: pl.Float32,
+        JobGroupLabels.other: pl.Float32,
         Cols.temp_location_sum: pl.Int64,
     }
     test_evaluation_expr_schema = {
